renamed queryItem to queryQuantity
