+++ src/index.js
@@ -5,7 +5,7 @@ const dynasty = require('dynasty')({});
 const pluralize = require('pluralize');
 const languageStrings = require('./languageStrings');
 const setQuantity = require('./intentDelegates/setQuantity');
-const queryItem = require('./intentDelegates/queryItem');
+const queryQuantity = require('./intentDelegates/queryQuantity');
 const removeItem = require('./intentDelegates/removeItem');
 const queryCreateTime = require('./intentDelegates/queryCreateTime');
 const incrementQuantity = require('./intentDelegates/incrementQuantity')
@@ -22,7 +22,7 @@ const handlers = {
 
     'RemoveItem': function() { removeItem(this, stewardItems) },
 
-    'QueryItem': function() { queryItem(this, stewardItems) },
+    'QueryQuantity': function() { queryQuantity(this, stewardItems) },
 
 	'QueryCreateTime' : function() { queryCreateTime(this, stewardItems) },
