+++ src/intentDelegates/queryQuantity.js
